+++ iOSEngineerCodeCheck/Models/Owner.swift
@@ -1,0 +1,10 @@
+//
+//  Owner.swift
+//  iOSEngineerCodeCheck
+//
+//  Created by TanakaHirokazu on 2021/10/30.
+//
+
+struct Owner: Decodable {
+    let avatarUrl: String?
+}
